Move CTagIndexLookup to iterators_ffi
