+++ src/redisearch_rs/rqe_iterators/Cargo.toml
@@ -44,6 +44,7 @@ workspace_hack.workspace = true
 approx.workspace = true
 decorum.workspace = true
 insta.workspace = true
+iterators_ffi = { path = "../c_entrypoint/iterators_ffi" }
 pretty_assertions.workspace = true
 proptest = { workspace = true, features = ["std"] }
 query_term_ffi = { path = "../c_entrypoint/query_term_ffi" }
