+++ src/redisearch_rs/rqe_iterators/src/inverted_index/tag.rs
@@ -9,13 +9,10 @@
 
 use std::ptr::NonNull;
 
-use ffi::{RedisSearchCtx, TagIndex};
+use ffi::RedisSearchCtx;
 use index_result::{RSIndexResult, RSOffsetSlice};
 use index_spec::IndexSpecReadGuard;
-use inverted_index::{
-    DecodedBy, DocIdsDecoder, IndexReader, IndexReaderCore, RawIndexReaderCore,
-    opaque::OpaqueEncoding,
-};
+use inverted_index::{DecodedBy, DocIdsDecoder, IndexReader, IndexReaderCore, RawIndexReaderCore};
 use query_term::RSQueryTerm;
 use ref_mode::{Active, Ref};
 use rqe_core::{DocId, RS_FIELDMASK_ALL};
@@ -33,49 +30,6 @@ use super::{InvIndIterator, core::RawInvIndIterator};
 pub trait TagLookup<E> {
     /// The inverted index currently stored for `tag`, if any.
     fn find(&self, tag: &[u8]) -> Option<&inverted_index::InvertedIndex<E>>;
-}
-
-/// [`TagLookup`] over the C `TagIndex`'s opaque `TrieMap` (`tag_index.values`).
-pub struct CTagIndexLookup(NonNull<TagIndex>);
-
-impl CTagIndexLookup {
-    /// Create a lookup over the given C [`TagIndex`].
-    ///
-    /// # Safety
-    ///
-    /// 1. `tag_index` must point to a valid [`TagIndex`] and remain valid for
-    ///    the lifetime of this lookup (and of any iterator holding it).
-    /// 2. `tag_index.values`, when non-null, must be a valid
-    ///    [`TrieMapOpaque`](trie_rs::TrieMapOpaque) pointer.
-    /// 3. The entries in `tag_index.values` must point to opaque
-    ///    [`InvertedIndex`](inverted_index::opaque::InvertedIndex)es whose
-    ///    encoding variant matches the `E` this lookup is used with.
-    pub const unsafe fn new(tag_index: NonNull<TagIndex>) -> Self {
-        Self(tag_index)
-    }
-}
-
-impl<E> TagLookup<E> for CTagIndexLookup
-where
-    E: OpaqueEncoding<Storage = inverted_index::InvertedIndex<E>>,
-{
-    fn find(&self, tag: &[u8]) -> Option<&inverted_index::InvertedIndex<E>> {
-        // SAFETY: 1. in `Self::new` guarantees `tag_index` is valid.
-        let tag_idx = unsafe { self.0.as_ref() };
-        if tag_idx.values.is_null() {
-            // No values trie means no postings for any tag.
-            return None;
-        }
-        // SAFETY: 2. in `Self::new` guarantees `values` is a valid `TrieMapOpaque`.
-        let trie = unsafe { &*tag_idx.values.cast::<trie_rs::TrieMapOpaque>() };
-        let idx = trie.find(tag)?;
-        // SAFETY: 3. in `Self::new` guarantees the trie entry points to a valid
-        // opaque `InvertedIndex`.
-        let opaque = idx.cast::<inverted_index::opaque::InvertedIndex>().as_ptr();
-        // SAFETY: 3. `from_opaque` panics when the encoding
-        // variant doesn't match `E`.
-        Some(E::from_opaque(unsafe { &*opaque }))
-    }
 }
 
 /// An iterator over documents matching a specific tag value, parameterised
@@ -96,23 +50,17 @@ where
 /// * `C` - The expiration checker type.
 /// * `L` - The [`TagLookup`] used to detect GC changes during revalidation.
 #[repr(C)]
-pub struct RawTag<
-    'query,
-    Rf: Ref,
-    E,
-    C = crate::expiration_checker::NoOpChecker,
-    L = CTagIndexLookup,
-> {
+pub struct RawTag<'query, Rf: Ref, E, L, C = crate::expiration_checker::NoOpChecker> {
     it: RawInvIndIterator<'query, Rf, RawIndexReaderCore<Rf, E>, C>,
     lookup: L,
 }
 
 /// Alias for an [`Active`] [`RawTag`] — the only instantiation with an
 /// [`RQEIterator`] impl today.
-pub type Tag<'index, E, C = crate::expiration_checker::NoOpChecker, L = CTagIndexLookup> =
-    RawTag<'index, Active<'index>, E, C, L>;
+pub type Tag<'index, E, L, C = crate::expiration_checker::NoOpChecker> =
+    RawTag<'index, Active<'index>, E, L, C>;
 
-impl<'index, E, C, L> Tag<'index, E, C, L>
+impl<'index, E, L, C> Tag<'index, E, L, C>
 where
     E: DecodedBy + 'index,
     <E as DecodedBy>::Decoder: DocIdsDecoder,
@@ -213,7 +161,7 @@ where
     }
 }
 
-impl<'index, E, C, L> RQEIterator<'index> for Tag<'index, E, C, L>
+impl<'index, E, L, C> RQEIterator<'index> for Tag<'index, E, L, C>
 where
     E: DecodedBy + 'index,
     <E as DecodedBy>::Decoder: DocIdsDecoder,
@@ -280,7 +228,7 @@ where
     }
 }
 
-impl<'index, E, C, L> ProfilePrint for Tag<'index, E, C, L>
+impl<'index, E, L, C> ProfilePrint for Tag<'index, E, L, C>
 where
     E: inverted_index::DecodedBy + 'index,
     <E as inverted_index::DecodedBy>::Decoder: inverted_index::DocIdsDecoder,
