+++ src/redisearch_rs/rqe_iterators/tests/integration/profile_print.rs
@@ -408,7 +408,7 @@ fn tag_with_query_term() {
         rqe_iterators::inverted_index::Tag::new(
             reader,
             mock_ctx.sctx(),
-            rqe_iterators::inverted_index::CTagIndexLookup::new(mock_ctx.tag_index()),
+            iterators_ffi::inverted_index::CTagIndexLookup::new(mock_ctx.tag_index()),
             RSQueryTerm::new("my_tag", 0, 0),
             0.0,
             rqe_iterators::NoOpChecker,
