+++ src/redisearch_rs/rqe_iterators_bencher/src/benchers/inverted_index/tag.rs
@@ -14,10 +14,8 @@ use std::hint::black_box;
 use criterion::{BenchmarkGroup, Criterion, measurement::Measurement};
 use index_result::RSQueryTerm;
 use inverted_index::{doc_ids_only::DocIdsOnly, opaque::OpaqueEncoding};
-use rqe_iterators::{
-    RQEIterator, SkipToOutcome,
-    inverted_index::{CTagIndexLookup, Tag},
-};
+use iterators_ffi::inverted_index::CTagIndexLookup;
+use rqe_iterators::{RQEIterator, SkipToOutcome, inverted_index::Tag};
 use rqe_iterators_test_utils::TestContext;
 
 use super::{INDEX_SIZE, SKIP_TO_STEP, SPARSE_DELTA, benchmark_group};
